+++ rpc/doc/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.21] - 2023-06-12
+
+- Update handlers to use new relay payload types
+
 ## [0.0.0.20] - 2023-06-01
 
 - Removed AAT type from rpc handlers