+++ runtime/docs/CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.41] - 2023-06-06
+## [0.0.0.41] - 2023-06-12
 
-- Add a new ChainConfig configuration field to servicer config
+- Add a new ServiceConfig configuration field to servicer config
 
 ## [0.0.0.40] - 2023-06-01